+++ niryo_robot_database/src/niryo_robot_database/api/__init__.py
@@ -1,1 +1,0 @@
-from .database_ros_wrapper import DatabaseRosWrapper