+++ niryo_robot_reports/src/niryo_robot_reports/DailyReportHandler.py
@@ -1,3 +1,5 @@
+from uuid import uuid4
+
 import rospy
 from io import StringIO
 from datetime import date, datetime
@@ -6,21 +8,18 @@ from niryo_robot_metrics.TuptimeWrapper import TuptimeWrapper
 
 from niryo_robot_reports.DailyReport import DailyReport
 from niryo_robot_reports.CloudAPI import MicroServiceError
+from niryo_robot_system_api_client import system_api_client
 
 # msg
 from std_msgs.msg import Empty
-from niryo_robot_msgs.msg import CommandStatus
 from niryo_robot_status.msg import RobotStatus
 
 
 class DailyReportHandler:
 
-    def __init__(self, cloud_api, reports_path, add_report_db, rm_report_db, get_all_files_paths_db):
+    def __init__(self, cloud_api, reports_path):
         self.__cloud_api = cloud_api
         self.__reports_path = reports_path
-        self.__add_report_db = add_report_db
-        self.__rm_report_db = rm_report_db
-        self.__get_all_files_paths_db = get_all_files_paths_db
         self.__current_date = str(date.today())
 
         self.__tuptime_wrapper = TuptimeWrapper()
@@ -29,8 +28,13 @@ class DailyReportHandler:
         report_path = '{}/{}'.format(self.__reports_path, report_name)
         self.__daily_report = DailyReport(report_path)
         self.__daily_report.set_date(self.__current_date)
-        add_report_response = self.__add_report_db('daily_report', report_name, report_path)
-        self.__current_id = add_report_response.message
+
+        add_report_response = system_api_client.add_file_path('daily_report', report_name, report_path)
+        if add_report_response.success:
+            self.__current_id = add_report_response.data['id']
+        else:
+            rospy.logerr('Unable to add the daily report to the database: {}'.format(add_report_response.detail))
+            self.__current_id = uuid4()
 
         rospy.Subscriber('~new_day', Empty, self.__new_day_callback)
         rospy.Subscriber('/niryo_robot_status/robot_status', RobotStatus, self.__robot_status_callback)
@@ -46,7 +50,7 @@ class DailyReportHandler:
         try:
             self.__cloud_api.daily_reports.send({'date': self.__current_date, 'report': self.__daily_report.content})
             self.__daily_report.delete()
-            self.__rm_report_db(self.__current_id)
+            system_api_client.rm_file_path(_id=self.__current_id)
         except MicroServiceError as microservice_error:
             rospy.logerr(str(microservice_error))
 
@@ -70,22 +74,22 @@ class DailyReportHandler:
         self.__daily_report.add_log(formatted_log, 'ROS', str(datetime.now()))
 
     def __send_failed_daily_reports(self):
-        daily_reports_response = self.__get_all_files_paths_db('daily_report')
-        if daily_reports_response.status == CommandStatus.SUCCESS:
-            for report in daily_reports_response.filepaths:
-                if report.date == self.__current_date:
+        daily_reports_response = system_api_client.get_file_paths('daily_report')
+        if daily_reports_response.success:
+            for report in daily_reports_response.data:
+                if report['date'] == self.__current_date:
                     continue
-                report_handler = DailyReport(report.path)
-                rospy.loginfo('Sending the report of {}'.format(report.date))
+                report_handler = DailyReport(report['path'])
+                rospy.loginfo('Sending the report of {}'.format(report['date']))
                 try:
                     self.__cloud_api.daily_reports.send(report_handler.content)
                 except MicroServiceError as microservice_error:
                     rospy.logerr(str(microservice_error))
-                    if (date.today() - datetime.strptime(report.date, '%Y-%m-%d').date()).days > 2:
-                        rospy.loginfo('Deleting the outdated daily report of {}'.format(report.date))
+                    if (date.today() - date.fromisoformat(report['date'])).days > 2:
+                        rospy.loginfo('Deleting the outdated daily report of {}'.format(report['date']))
                         report_handler.delete()
-                        self.__rm_report_db(report.id)
+                        system_api_client.rm_file_path(report['id'])
                     continue
 
                 report_handler.delete()
-                self.__rm_report_db(report.id)
+                system_api_client.rm_file_path(report['id'])