+++ niryo_robot_reports/src/niryo_robot_reports/TestReportHandler.py
@@ -1,24 +1,20 @@
 import json
 import rospy
-from datetime import datetime
-from dateutil import parser
+from datetime import datetime, date
 
 from niryo_robot_reports.TestReport import TestReport
 from niryo_robot_reports.CloudAPI import MicroServiceError
+from niryo_robot_system_api_client import system_api_client
 
 # msg
 from std_msgs.msg import String
-from niryo_robot_msgs.msg import CommandStatus
 
 
 class TestReportHandler:
 
-    def __init__(self, cloud_api, reports_path, add_report_db, rm_report_db, get_all_files_paths_db):
+    def __init__(self, cloud_api, reports_path):
         self.__cloud_api = cloud_api
         self.__reports_path = reports_path
-        self.__add_report_db = add_report_db
-        self.__rm_report_db = rm_report_db
-        self.__get_all_files_paths_db = get_all_files_paths_db
 
         self.__send_failed_test_reports()
 
@@ -40,22 +36,22 @@ class TestReportHandler:
             report_path = '{}/{}'.format(self.__reports_path, report_name)
             report_handler = TestReport(report_path)
             report_handler.set_content(parsed_json)
-            self.__add_report_db('test_report', report_name, report_path)
+            system_api_client.add_file_path('test_report', report_name, report_path)
 
     def __send_failed_test_reports(self):
-        test_reports_response = self.__get_all_files_paths_db('test_report')
-        if test_reports_response.status == CommandStatus.SUCCESS:
-            for report in test_reports_response.filepaths:
-                report_handler = TestReport(report.path)
+        test_reports_response = system_api_client.get_file_paths('test_report')
+        if test_reports_response.success:
+            for report in test_reports_response.data:
+                report_handler = TestReport(report['path'])
                 rospy.loginfo('Sending the test report of {}'.format(report_handler.content['date']))
 
                 try:
                     self.__cloud_api.test_reports.send(report_handler.content)
                     report_handler.delete()
-                    self.__rm_report_db(report.id)
+                    system_api_client.rm_file_path(report['id'])
                 except MicroServiceError as microservice_error:
                     rospy.logerr(str(microservice_error))
-                    if (datetime.now() - parser.parse(report_handler.content['date'])).days > 2:
+                    if (date.today() - date.fromisoformat(report['date'])).days > 2:
                         rospy.loginfo('Deleting the outdated test report of {}'.format(report_handler.content['date']))
                         report_handler.delete()
-                        self.__rm_report_db(report.id)
+                        system_api_client.rm_file_path(report['id'])