+++ niryo_robot_system_api_client/src/niryo_robot_system_api_client/api/__init__.py
@@ -1,0 +1,1 @@
+from .system_api_ros_wrapper import SystemAPIRosWrapper